% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fuentes_datos}
\alias{fuentes_datos}
\title{Registro de fuentes de los datos}
\format{
Un tibble con columnas \code{fuente}, \code{detalle}, \code{version}, \code{url},
\code{fecha} y \code{md5}.
}
\usage{
fuentes_datos
}
\description{
Registro de fuentes de los datos
}
\keyword{datasets}
