% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hpo_genes}
\alias{hpo_genes}
\title{Fenotipos HPO de los genes del panel}
\format{
Un tibble con columnas \code{gen}, \code{hpo_id} y \code{hpo_nombre}.
}
\source{
Human Phenotype Ontology, archivo genes_to_phenotype.txt,
\url{https://hpo.jax.org}.
}
\usage{
hpo_genes
}
\description{
Fenotipos HPO de los genes del panel
}
\keyword{datasets}
