% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anotar.R
\name{iei_limpiar_cache}
\alias{iei_limpiar_cache}
\title{Borrar la cache de anotaciones}
\usage{
iei_limpiar_cache(cache = tools::R_user_dir("ieiprio", "cache"))
}
\arguments{
\item{cache}{Carpeta de la cache.}
}
\description{
Borrar la cache de anotaciones
}
