% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorizar.R
\name{iei_pesos}
\alias{iei_pesos}
\title{Pesos por defecto de la priorizacion}
\usage{
iei_pesos()
}
\value{
Una lista con nombres.
}
\description{
Devuelve la lista de puntos que usa \code{\link[=iei_priorizar]{iei_priorizar()}}. Puedes modificar
cualquier valor y pasar la lista en el argumento \code{pesos}.
}
\examples{
p <- iei_pesos()
p$clinvar_patogenica <- 5
}
