% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iuis_alias}
\alias{iuis_alias}
\title{Alias y simbolos anteriores de los genes del panel}
\format{
Un tibble con columnas \code{alias} y \code{gen} (simbolo oficial).
}
\source{
PanelApp, que a su vez toma los datos de HGNC.
}
\usage{
iuis_alias
}
\description{
Alias y simbolos anteriores de los genes del panel
}
\keyword{datasets}
