% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iuis_categorias}
\alias{iuis_categorias}
\title{Categorias IUIS de los genes del panel}
\format{
Un tibble con columnas \code{gen}, \code{categoria_iuis} (1 a 10) y
\code{categoria_nombre}.
}
\source{
Clasificacion IUIS, Journal of Clinical Immunology.
}
\usage{
iuis_categorias
}
\description{
Curacion manual de la clasificacion IUIS de errores innatos de la
inmunidad. Un gen puede estar en mas de una categoria.
}
\keyword{datasets}
