% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iuis_panel}
\alias{iuis_panel}
\title{Panel de genes de errores innatos de la inmunidad (IUIS)}
\format{
Un tibble con una fila por combinacion de gen y enfermedad:
\describe{
\item{gen}{Simbolo oficial HGNC.}
\item{categoria_iuis}{Categoria IUIS, entero de 1 a 10.}
\item{categoria_nombre}{Nombre de la categoria.}
\item{enfermedad}{Enfermedad asociada.}
\item{herencia}{Modo de herencia: AR, AD, XL, XLR o AR/AD.}
\item{ganancia_funcion}{\code{TRUE} si el mecanismo es ganancia de funcion.}
}
}
\source{
Clasificacion IUIS de errores innatos de la inmunidad,
Journal of Clinical Immunology.
}
\usage{
iuis_panel
}
\description{
Genes asociados a errores innatos de la inmunidad segun la clasificacion
de la International Union of Immunological Societies (IUIS). La version
actual es un panel semilla para desarrollo; se reemplazara por la
clasificacion IUIS 2024 completa.
}
\details{
Un gen puede aparecer en mas de una fila si causa enfermedades distintas
segun el modo de herencia o el mecanismo (por ejemplo STAT1).
}
\keyword{datasets}
