% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ieiprio-package.R
\docType{package}
\name{ieiprio-package}
\alias{ieiprio}
\alias{ieiprio-package}
\title{ieiprio: Priorizacion de Variantes en Genes de Errores Innatos de la Inmunidad}
\description{
Lee archivos VCF, filtra variantes en genes de errores innatos de la inmunidad segun la clasificacion IUIS, las anota con Ensembl VEP y las prioriza segun modo de herencia y fenotipo clinico, generando un reporte por paciente. Herramienta de apoyo e investigacion, no de diagnostico.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/TU_USUARIO/ieiprio}
  \item Report bugs at \url{https://github.com/TU_USUARIO/ieiprio/issues}
}

}
\author{
\strong{Maintainer}: Tu Nombre \email{tu.correo@ejemplo.com}

Authors:
\itemize{
  \item Tu Nombre \email{tu.correo@ejemplo.com}
}

}
\keyword{internal}
