% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iuis_alias}
\alias{iuis_alias}
\title{Alias de simbolos de genes del panel}
\format{
Un tibble con dos columnas:
\describe{
\item{alias}{Nombre alternativo o anterior.}
\item{gen}{Simbolo oficial HGNC.}
}
}
\source{
HUGO Gene Nomenclature Committee, \url{https://www.genenames.org}.
}
\usage{
iuis_alias
}
\description{
Tabla para traducir nombres antiguos o alternativos al simbolo oficial
HGNC usado en \link{iuis_panel}.
}
\keyword{datasets}
